+++ transport/tchannel/channel_transport.go
@@ -21,7 +21,7 @@
 package tchannel
 
 import (
-	"fmt"
+	"errors"
 
 	"go.uber.org/yarpc/internal/sync"
 
@@ -29,19 +29,23 @@ import (
 	"github.com/uber/tchannel-go"
 )
 
+var errChannelOrServiceNameIsRequired = errors.New(
+	"cannot instantiate tchannel.ChannelTransport: " +
+		"please provide a service name with the ServiceName option " +
+		"or an existing Channel with the WithChannel option")
+
 // NewChannelTransport is a YARPC transport that facilitates sending and
 // receiving YARPC requests through TChannel. It uses a shared TChannel
 // Channel for both, incoming and outgoing requests, ensuring reuse of
 // connections and other resources.
 //
 // Either the local service name (with the ServiceName option) or a user-owned
-// TChannel (with the WithChannel option) MUST be specified or this transport
-// will fail to Start.
+// TChannel (with the WithChannel option) MUST be specified.
 //
 // ChannelTransport uses the underlying TChannel Channel for load balancing
 // and peer managament. A future version of YARPC will include support for
 // peer.Chooser-based TChannel transports.
-func NewChannelTransport(opts ...TransportOption) *ChannelTransport {
+func NewChannelTransport(opts ...TransportOption) (*ChannelTransport, error) {
 	var config transportConfig
 	config.tracer = opentracing.GlobalTracer()
 	for _, opt := range opts {
@@ -53,22 +57,21 @@ func NewChannelTransport(opts ...TransportOption) *ChannelTransport {
 	// an error return.
 	var err error
 	ch := config.ch
-	if config.ch == nil {
+
+	if ch == nil {
 		if config.name == "" {
-			err = fmt.Errorf("can't instantiate TChannelChannelTransport without channel or service name option")
+			err = errChannelOrServiceNameIsRequired
 		} else {
-			ch, err = tchannel.NewChannel(config.name, &tchannel.ChannelOptions{
-				Tracer: config.tracer,
-			})
+			opts := tchannel.ChannelOptions{Tracer: config.tracer}
+			ch, err = tchannel.NewChannel(config.name, &opts)
 		}
 	}
 
 	return &ChannelTransport{
 		ch:     ch,
-		err:    err,
 		addr:   config.addr,
 		tracer: config.tracer,
-	}
+	}, err
 }
 
 // ChannelTransport maintains TChannel peers and creates inbounds and outbounds for
@@ -76,7 +79,6 @@ func NewChannelTransport(opts ...TransportOption) *ChannelTransport {
 type ChannelTransport struct {
 	ch     Channel
 	name   string
-	err    error
 	addr   string
 	tracer opentracing.Tracer
 
@@ -101,11 +103,6 @@ func (t *ChannelTransport) Start() error {
 }
 
 func (t *ChannelTransport) start() error {
-	// Return error deferred from constructor for the construction of a TChannel.
-	if t.err != nil {
-		return t.err
-	}
-
 	if t.ch.State() == tchannel.ChannelListening {
 		// Channel.Start() was called before RPC.Start(). We still want to
 		// update the Handler and what t.addr means, but nothing else.